first attempts to separate mpi4py-wrappers from communication objects in DNDarrays
