+++ heat/__init__.py
@@ -18,3 +18,4 @@ from . import sparse
 from . import spatial
 from . import utils
 from . import preprocessing
+from . import communication_backends
