+++ heat/communication_backends/__init__.py
@@ -1,0 +1,5 @@
+"""
+Add the communication_backends functions to the ht.communication namespace
+"""
+
+from .mpi4py4torch import *
